+++ D_S_and_A/Algorithms/c_max_heap_sort.inl
@@ -7,7 +7,7 @@ template<class MyType, template <class> class NodeType>
 std::vector<MyType>& c_max_heap_sort<MyType, NodeType>::sort(std::vector<MyType>& arr)
 {
 	c_max_heap<MyType, NodeType> heap{ arr[0] };
-	for (int i{ 1 }; i < arr.size(); ++i)
+	for (int i{ 1 }; i < static_cast<int>(arr.size()); ++i)
 		heap.add_item(arr[i]);
 	for (int i{ static_cast<int>(arr.size()) - 1 }; i >= 0; --i)//go backward through array adding largest
 	{
